resturcture the code into modules
