+++ grpc-calculator/Cargo.toml
@@ -8,6 +8,7 @@ tonic = "0.11"
 tokio = { version = "1", features = ["full"]   }
 prost = "0.12"
 tonic-reflection = "0.11"
+tonic-health = "0.12.0"
 
 [build-dependencies]
 tonic-build = "0.11"
