+++ grpc-calculator/src/main.rs
@@ -1,5 +1,8 @@
-use proto::calculator_server::{Calculator, CalculatorServer};
+pub mod service;
+
+use service::calculator_service;
 use tonic::transport::Server;
+use tonic_health::server::HealthReporter;
 use tonic_reflection;
 
 mod proto {
@@ -7,62 +10,20 @@ mod proto {
 
     pub(crate) const FILE_DESCRIPTOR_SET: &[u8] =
         tonic::include_file_descriptor_set!("calculator_descriptor");
-
-}
-
-#[derive(Debug, Default)]
-struct CalculatorService {}
-
-#[tonic::async_trait]
-impl Calculator for CalculatorService {
-    async fn add(
-        &self,
-        request: tonic::Request<proto::CalculationRequest>,
-    ) -> Result<tonic::Response<proto::CalculationResponse>, tonic::Status> {
-        println!("Got a request: {:?}", request);
-
-        let input = request.get_ref();
-        let response = proto::CalculationResponse {
-            result: input.a + input.b,
-        };
-
-        Ok(tonic::Response::new(response))
-    }
-
-    async fn divide(
-        &self,
-        request: tonic::Request<proto::CalculationRequest>,
-    ) -> Result<tonic::Response<proto::CalculationResponse>, tonic::Status> {
-        println!("Got a request: {:?}", request);
-
-        let input = request.get_ref();
-
-        if input.b == 0 {
-            return Err(tonic::Status::invalid_argument("Cannot Divide by 0!"))
-        }
-
-        let response = proto::CalculationResponse {
-            result: input.a / input.b,
-        };
-
-        Ok(tonic::Response::new(response))
-    }
-
 }
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
-    
     let addr = "[::1]:50051".parse()?;
-    let calc = CalculatorService::default();
-    
+    let calc = calculator_service::CalculatorService::default();
+
     let service = tonic_reflection::server::Builder::configure()
         .register_encoded_file_descriptor_set(proto::FILE_DESCRIPTOR_SET)
         .build()?;
 
     Server::builder()
         .add_service(service)
-        .add_service(CalculatorServer::new(calc))
+        .add_service(calculator_service::CalculatorServer::new(calc))
         .serve(addr)
         .await?;
