+++ grpc-calculator/src/service/mod.rs
@@ -1,0 +1,1 @@
+pub mod calculator_service;
